Add files via upload
The minimum bounding rectangle which represents the Chicago region, might not be the best way of representing the data since the study are should be defined well to analyze the regional pattern distribution. This can give rise to problems in terms of regional interpretation where the boundaries of the Chicago state is not correctly represented and can cause confusion and make it difficult to understand any edge effects, scale effects and so on. It would be better to define the study area as the boundary of the region instead of only considering crime locations. This would help to understand any interactions with other cities, make an aerial conclusion